requirements.txt
pyspark
python-dotenv
mlflow
kaggle
psutil
matplotlib
seaborn